+++ test/att_db/att_db_util_test.c
@@ -55,20 +55,7 @@
 #include "btstack_util.h"
 #include "bluetooth.h"
  
-#include "le_counter.h"
-
-#if 0
-PRIMARY_SERVICE, GAP_SERVICE
-CHARACTERISTIC, GAP_DEVICE_NAME, READ, "SPP+LE Counter"
-
-PRIMARY_SERVICE, GATT_SERVICE
-CHARACTERISTIC, GATT_SERVICE_CHANGED, READ,
-
-// Counter Service
-PRIMARY_SERVICE, 0000FF10-0000-1000-8000-00805F9B34FB
-// Counter Characteristic, with read and notify
-CHARACTERISTIC,  0000FF11-0000-1000-8000-00805F9B34FB, READ | NOTIFY | DYNAMIC,
-#endif
+#include "att_db_util_test.h"
 
 // 0000FF10-0000-1000-8000-00805F9B34FB
 uint8_t counter_service_uuid[] = { 0x00, 0x00, 0xFF, 0x10, 0x00, 0x00, 0x10, 0x00, 0x80, 0x00, 0x00, 0x80, 0x5F, 0x9B, 0x34, 0xFB};
@@ -77,7 +64,7 @@ uint8_t counter_characteristic_uuid[] = { 0x00, 0x00, 0xFF, 0x11, 0x00, 0x00, 0x
 
 void CHECK_EQUAL_ARRAY(const uint8_t * expected, uint8_t * actual, int size){
     for (int i=0; i<size; i++){
-        // printf("%03u: %02x - %02x\n", i, expected[i], actual[i]);
+        printf("%03u: %02x - %02x\n", i, expected[i], actual[i]);
         BYTES_EQUAL(expected[i], actual[i]);
     }
 }
@@ -96,16 +83,18 @@ TEST_GROUP(AttDbUtil){
 
 TEST(AttDbUtil, LeCounterDb){
     att_db_util_add_service_uuid16(GAP_SERVICE_UUID);
-    att_db_util_add_characteristic_uuid16(GAP_DEVICE_NAME_UUID, ATT_PROPERTY_READ, (uint8_t*)"SPP+LE Counter", 14);
+    att_db_util_add_characteristic_uuid16(GAP_DEVICE_NAME_UUID, ATT_PROPERTY_READ, ATT_SECURITY_NONE, ATT_SECURITY_NONE, (uint8_t*)"SPP+LE Counter", 14);
 
     att_db_util_add_service_uuid16(0x1801);
-    att_db_util_add_characteristic_uuid16(0x2a05, ATT_PROPERTY_READ, NULL, 0);
+    att_db_util_add_characteristic_uuid16(0x2a05, ATT_PROPERTY_READ, ATT_SECURITY_NONE, ATT_SECURITY_NONE, NULL, 0);
 
     att_db_util_add_service_uuid128(counter_service_uuid);
-    att_db_util_add_characteristic_uuid128(counter_characteristic_uuid, ATT_PROPERTY_READ | ATT_PROPERTY_NOTIFY | ATT_PROPERTY_DYNAMIC, NULL, 0);
+    att_db_util_add_characteristic_uuid128(counter_characteristic_uuid, ATT_PROPERTY_READ | ATT_PROPERTY_NOTIFY | ATT_PROPERTY_DYNAMIC, ATT_SECURITY_NONE, ATT_SECURITY_NONE, NULL, 0);
 
     uint8_t * addr = att_db_util_get_address();
     uint16_t  size = att_db_util_get_size();
+
+    printf_hexdump(addr, size);
 
     CHECK_EQUAL(size, (uint16_t)sizeof(profile_data));
     CHECK_EQUAL_ARRAY(profile_data, addr, size);